test: migrate RevisionFinder tests in source
